refactor(library): move meeting speaker rename into Library/meeting_speaker_rename.py
Pure relocation ahead of TASK-31745: the after-the-fact rename data layer
(`can_rename_meeting_speakers`, `rename_meeting_speaker`, the refusal
constants, and the transcript read/render/write helpers) moves out of
`library_media_canvas.py` UNCHANGED so a second surface -- the live media
reader -- can use it without importing a widget module. The canvas
re-exports every moved name, so existing importers are untouched.

One test line does change: the redaction test monkeypatches
`update_meeting_json` where `rename_meeting_speaker` looks it up, which is
now the new module. Left on the canvas it would have patched a name nobody
reads and passed vacuously.
